==> 03_notebooks/predictive_notebooks/1_dataset_download.ipynb <==
## Data extraction

This notebook is downloading the already created dataset for AI in Automotive topic from kaggle.

The dataset includes 25K patent and paper based documents with their titles, abstracts, and corresponding publication or filing dates.


import os
import zipfile
import shutil

DATASET = "mertasaroglu/nasa-trl-labeling-ia-in-automotive-sector"


ROOT = os.path.abspath("../..")  
TARGET_DIR = os.path.join(ROOT, "01_data", "predictive_model")

os.makedirs(TARGET_DIR, exist_ok=True)


ZIP_NAME = "nasa-trl-labeling-ia-in-automotive-sector.zip"

print(">> Downloading Kaggle dataset...")
!kaggle datasets download -d {DATASET} -p "{TARGET_DIR}"


zip_path = os.path.join(TARGET_DIR, ZIP_NAME)


print(">> Extracting ZIP...")
with zipfile.ZipFile(zip_path, "r") as z:
    z.extractall(TARGET_DIR)


os.remove(zip_path)

print(">> Done. Extracted files:")
print(os.listdir(TARGET_DIR))
